Move src/lib/client to src/client
