+++ web/src/InstallationFinished.test.jsx
@@ -24,11 +24,11 @@ import React from "react";
 import { screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import { authRender } from "./test-utils";
-import { createClient } from "./lib/client";
+import { createClient } from "./client";
 
 import InstallationFinished from "./InstallationFinished";
 
-jest.mock("./lib/client");
+jest.mock("./client");
 
 let startProbingFn = jest.fn();
 let rebootSystemFn = jest.fn();
