+++ web/src/RootPassword.test.jsx
@@ -24,11 +24,11 @@ import React from "react";
 import { screen, waitFor, within } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import { authRender } from "./test-utils";
-import { createClient } from "./lib/client";
+import { createClient } from "./client";
 
 import RootPassword from "./RootPassword";
 
-jest.mock("./lib/client");
+jest.mock("./client");
 
 let isRootPasswordSetFn;
 let setRootPasswordFn = jest.fn();
